Fix to work w/constants (and subrefs in the stash)
This module dies with ‘Not a GLOB reference’ when encountering a
scalar reference in a stash, which is how constants have been stored
since 5.10 or so, because it assumes that all stash elements contain
typeglobs.

This commit changes it to account for anything that is not a typeglob
already by accessing it via *{$name} syntax.

This also makes the module compatible with the CV-
in-stash optimization that will likely be in perl
5.28.  For details on this optimization, see:
https://rt.perl.org/Ticket/Display.html?id=132252#txn-1500037